integration_projects/requirements.txt
asyncio==3.4.3
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
envyaml==1.10.211231
exceptiongroup==1.1.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
isort==5.12.0
lxml==4.9.2
mysql-connector==2.2.9
mysql-connector-python==8.0.33
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pep8==1.7.1
pip-autoremove==0.10.0
pip-review==1.3.0
platformdirs==3.5.0
pluggy==1.0.0
protobuf==3.20.3
PyJWT==2.7.0
PyMySQL==1.0.3
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
simple-salesforce==1.12.3
six==1.16.0
SQLAlchemy==2.0.13
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
zeep==4.2.1
